+++ fxgrpcserver/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.2.0](https://github.com/ankorstore/yokai/compare/fxgrpcserver/v1.1.0...fxgrpcserver/v1.2.0) (2024-04-05)
+
+
+### Features
+
+* **fxgrpcserver:** Added support for listener address ([#204](https://github.com/ankorstore/yokai/issues/204)) ([4bf8968](https://github.com/ankorstore/yokai/commit/4bf89684121dae4e24bbc63a4199102da378b1ce))
+
 ## [1.1.0](https://github.com/ankorstore/yokai/compare/fxgrpcserver/v1.0.0...fxgrpcserver/v1.1.0) (2024-03-15)